chatbot/requirements.txt
tensorflow==2.3.1
python-telegram-bot
